unitest

+++ Libft/libft.h
@@ -14,7 +14,8 @@
 # define LIBFT_H
 
 #include <stdlib.h>
-#include <stdio.h>
+#include <unistd.h>
+
 int	ft_isalpha(int c);
 int	ft_isdigit(int c);
 int	isalnum(int c);
@@ -22,7 +23,7 @@ int	ft_isascii(int c);
 int	ft_isprint(int c);
 int	ft_toupper(int c);
 int	ft_tolower(int c);
-int	ft_strlen(const char *s);
+size_t  ft_strlen(const char *s);
 int	ft_atoi(const char *str);
 int	ft_strncmp(const char *s1, const char *s2, size_t n);
 int	ft_memcmp(const void *s1, const void *s2, size_t n);
